Known high-risk type groups:

+++ src/telegram_bot/types/pre_checkout_query.cr
@@ -5,7 +5,7 @@ module TelegramBot
     property id : String
     property from : User
     property currency : String
-    property total_amount : Int32
+    property total_amount : Int64
     property invoice_payload : String
     property shipping_option_id : String?
     property order_info : OrderInfo?
